Put Stark operations in their own file.
