+++ evm/src/arithmetic/arithmetic_stark.rs
@@ -1,6 +1,5 @@
 use std::marker::PhantomData;
 
-use ethereum_types::U256;
 use itertools::Itertools;
 use plonky2::field::extension::{Extendable, FieldExtension};
 use plonky2::field::packed::PackedField;
@@ -8,6 +7,7 @@ use plonky2::field::polynomial::PolynomialValues;
 use plonky2::hash::hash_types::RichField;
 use plonky2::util::transpose;
 
+use crate::arithmetic::operations::Operation;
 use crate::arithmetic::{add, columns, compare, modular, mul, sub};
 use crate::constraint_consumer::{ConstraintConsumer, RecursiveConstraintConsumer};
 use crate::lookup::{eval_lookups, eval_lookups_circuit, permuted_cols};
@@ -15,174 +15,9 @@ use crate::permutation::PermutationPair;
 use crate::stark::Stark;
 use crate::vars::{StarkEvaluationTargets, StarkEvaluationVars};
 
-#[inline]
-fn u64_to_array<F: RichField>(out: &mut [F], x: u64) {
-    debug_assert!(out.len() == 4);
-
-    const MASK: u64 = (1 << 16) - 1;
-    out[0] = F::from_canonical_u64(x & MASK);
-    out[1] = F::from_canonical_u64((x >> 16) % MASK);
-    out[2] = F::from_canonical_u64((x >> 32) % MASK);
-    out[3] = F::from_canonical_u64((x >> 48) % MASK);
-}
-
-fn u256_to_array<F: RichField>(out: &mut [F], x: U256) {
-    debug_assert!(out.len() == columns::N_LIMBS);
-
-    u64_to_array(&mut out[0..4], x.0[0]);
-    u64_to_array(&mut out[4..8], x.0[1]);
-    u64_to_array(&mut out[8..12], x.0[2]);
-    u64_to_array(&mut out[12..16], x.0[3]);
-}
-
 #[derive(Copy, Clone)]
 pub struct ArithmeticStark<F, const D: usize> {
     pub f: PhantomData<F>,
-}
-
-pub trait Operation<F: RichField> {
-    /// Convert operation into one or two rows of the trace.
-    ///
-    /// Morally these types should be [F; columns::NUM_ARITH_COLUMNS], but we
-    /// use vectors because that's what utils::transpose expects.
-    fn to_rows(&self) -> (Vec<F>, Option<Vec<F>>);
-}
-
-struct SimpleBinaryOp {
-    op: usize,
-    input0: U256,
-    input1: U256,
-}
-
-impl SimpleBinaryOp {
-    pub fn new(op: usize, input0: U256, input1: U256) -> Self {
-        assert!(
-            op == columns::IS_ADD
-                || op == columns::IS_SUB
-                || op == columns::IS_MUL
-                || op == columns::IS_LT
-                || op == columns::IS_GT
-        );
-        Self { op, input0, input1 }
-    }
-}
-
-impl<F: RichField> Operation<F> for SimpleBinaryOp {
-    fn to_rows(&self) -> (Vec<F>, Option<Vec<F>>) {
-        let mut row = vec![F::ZERO; columns::NUM_ARITH_COLUMNS];
-        row[self.op] = F::ONE;
-
-        // Each of these operations uses the same columns for input; the
-        // asserts ensure no-one changes this.
-        debug_assert!([
-            columns::ADD_INPUT_0,
-            columns::SUB_INPUT_0,
-            columns::MUL_INPUT_0,
-            columns::CMP_INPUT_0,
-        ]
-        .iter()
-        .all(|x| *x == columns::GENERAL_INPUT_0));
-        debug_assert!([
-            columns::ADD_INPUT_1,
-            columns::SUB_INPUT_1,
-            columns::MUL_INPUT_1,
-            columns::CMP_INPUT_1,
-        ]
-        .iter()
-        .all(|x| *x == columns::GENERAL_INPUT_1));
-
-        u256_to_array(&mut row[columns::GENERAL_INPUT_0], self.input0);
-        u256_to_array(&mut row[columns::GENERAL_INPUT_1], self.input1);
-
-        // This is ugly, but it avoids the huge amount of boilderplate
-        // required to dispatch directly to each add/sub/etc. operation.
-        match self.op {
-            columns::IS_ADD => add::generate(&mut row),
-            columns::IS_SUB => sub::generate(&mut row),
-            columns::IS_MUL => mul::generate(&mut row),
-            columns::IS_LT | columns::IS_GT => compare::generate(&mut row, self.op),
-            _ => panic!("unrecognised operation"),
-        }
-
-        (row, None)
-    }
-}
-
-pub struct ModularBinaryOp {
-    op: usize,
-    input0: U256,
-    input1: U256,
-    modulus: U256,
-}
-
-impl ModularBinaryOp {
-    pub fn new(op: usize, input0: U256, input1: U256, modulus: U256) -> Self {
-        assert!(op == columns::IS_ADDMOD || op == columns::IS_SUBMOD || op == columns::IS_MULMOD);
-        Self {
-            op,
-            input0,
-            input1,
-            modulus,
-        }
-    }
-}
-
-fn modular_to_rows_helper<F: RichField>(
-    op: usize,
-    input0: U256,
-    input1: U256,
-    modulus: U256,
-) -> (Vec<F>, Option<Vec<F>>) {
-    let mut row1 = vec![F::ZERO; columns::NUM_ARITH_COLUMNS];
-    let mut row2 = vec![F::ZERO; columns::NUM_ARITH_COLUMNS];
-
-    row1[op] = F::ONE;
-
-    u256_to_array(&mut row1[columns::MODULAR_INPUT_0], input0);
-    u256_to_array(&mut row1[columns::MODULAR_INPUT_1], input1);
-    u256_to_array(&mut row1[columns::MODULAR_MODULUS], modulus);
-
-    modular::generate(&mut row1, &mut row2, op);
-
-    (row1, Some(row2))
-}
-
-impl<F: RichField> Operation<F> for ModularBinaryOp {
-    fn to_rows(&self) -> (Vec<F>, Option<Vec<F>>) {
-        modular_to_rows_helper(self.op, self.input0, self.input1, self.modulus)
-    }
-}
-
-pub struct ModOp {
-    input: U256,
-    modulus: U256,
-}
-
-impl<F: RichField> Operation<F> for ModOp {
-    fn to_rows(&self) -> (Vec<F>, Option<Vec<F>>) {
-        modular_to_rows_helper(columns::IS_MOD, self.input, U256::zero(), self.modulus)
-    }
-}
-
-pub struct DivOp {
-    numerator: U256,
-    denominator: U256,
-}
-
-impl<F: RichField> Operation<F> for DivOp {
-    fn to_rows(&self) -> (Vec<F>, Option<Vec<F>>) {
-        let mut row1 = vec![F::ZERO; columns::NUM_ARITH_COLUMNS];
-        let mut row2 = vec![F::ZERO; columns::NUM_ARITH_COLUMNS];
-
-        row1[columns::IS_DIV] = F::ONE;
-
-        u256_to_array(&mut row1[columns::DIV_NUMERATOR], self.numerator);
-        u256_to_array(&mut row1[columns::DIV_DENOMINATOR], self.denominator);
-
-        modular::generate(&mut row1, &mut row2, columns::IS_DIV);
-
-        (row1, Some(row2))
-    }
 }
 
 const RANGE_MAX: usize = 1usize << 16; // Range check strict upper bound
@@ -318,9 +153,8 @@ mod tests {
     use plonky2::field::types::{Field, PrimeField64};
     use plonky2::plonk::config::{GenericConfig, PoseidonGoldilocksConfig};
 
-    use super::{
-        columns, ArithmeticStark, DivOp, ModOp, ModularBinaryOp, Operation, SimpleBinaryOp,
-    };
+    use super::{columns, ArithmeticStark};
+    use crate::arithmetic::operations::*;
     use crate::stark_testing::{test_stark_circuit_constraints, test_stark_low_degree};
 
     #[test]
